back to uint32; delete fp

+++ sources/utils.cpp
@@ -86,7 +86,7 @@ string intToString(uint64_t num) {
   return s;
 }
 
-uint64_t revhash(uint64_t x) {
+uint32_t revhash(uint32_t x) {
 	x = ((x >> 16) ^ x) * 0x2c1b3c6d;
 	x = ((x >> 16) ^ x) * 0x297a2d39;
 	x = ((x >> 16) ^ x);
